+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+- Added: 'UDP.closed' property: readonly property that is True on closed UDP instances.
+- Added: It is not possible to call 'UDP.listen' on a udp instance that was already listening.
+- Added: 'UDP.send' returns -1 on closed instances.
+- Change: UDPReceiveCallback renamed to DatagramCallback.
+
 ## 1.0.2
 
 - Package Maintenance: description rewritten.